Reorganize JAX plugin code
