+++ dali/python/nvidia/dali/plugin/jax/__init__.py
@@ -11,8 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import sys
+import jax
 
+from distutils.version import LooseVersion
 from .iterator import DALIGenericIterator
+
+assert sys.version_info.major == 3 and sys.version_info.minor >= 8, \
+    "DALI JAX support requires Python 3.8 or above"
+
+
+assert LooseVersion(jax.__version__) >= LooseVersion('0.4.11'), \
+    "DALI JAX support requires JAX 0.4.11 or above"
 
 
 __all__ = ["DALIGenericIterator"]
